// src/api/orval/model/vpnServersDtoVpnServerDiscoveryDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { EnumsVpnServerDiscoveryStatus } from './enumsVpnServerDiscoveryStatus';
import type { EnumsVpnServerType } from './enumsVpnServerType';

export interface VpnServersDtoVpnServerDiscoveryDto {
  id?: number;
  serverType?: EnumsVpnServerType;
  /** @nullable */
  apiUrl?: string | null;
  /** @nullable */
  suggestedName?: string | null;
  /** @nullable */
  publicIp?: string | null;
  /** @nullable */
  version?: string | null;
  isEnableWss?: boolean;
  status?: EnumsVpnServerDiscoveryStatus;
  createdAt?: string;
  lastSeenUtc?: string;
  /** @nullable */
  resolvedVpnServerId?: number | null;
}