// src/api/orval/model/vpnServersResponsesVpnServerDiscoveriesResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { VpnServersDtoVpnServerDiscoveryDto } from './vpnServersDtoVpnServerDiscoveryDto';

export interface VpnServersResponsesVpnServerDiscoveriesResponse {
  /** @nullable */
  discoveries?: VpnServersDtoVpnServerDiscoveryDto[] | null;
}